requirements.txt
Flask==2.3.2
langchain==0.1.20
google-generativeai==0.5.4
tiktoken==0.5.1
python-dotenv==1.0.1
markdown==3.6
langchain_google_genai
sentence-transformers
python-docx
faiss-cpu
faiss-gpu